requirements.txt
beautifulsoup4
requests
jinja2
pyyaml
html5lib
dnslib
pysocks
requests[socks]
aiosocks